Release 2.0.1 version
- The .cpp versions of the core functions are no longer integrated into their .h files (they are now separated into .cpp and .h for later modification).

- Removed the future library for parallelization, returning to the old cluster parallelization on Windows to avoid library bloat.

- Author's update to the Description file.
