+++ NAMESPACE
@@ -1,12 +1,9 @@
 useDynLib(mutualinf, .registration=TRUE)
 importFrom(Rcpp, evalCpp)
-exportPattern("^[[:alpha:]]+")
 
 export(mutual)
 export(prepare_data)
 import(data.table)
 import(parallel)
-import(future)
-import(future.apply)
 import(runner)
 import(stats)
